+++ tests/Fixtures/app/AppKernel.php
@@ -6,6 +6,7 @@ use ApiPlatform\Symfony\Bundle\ApiPlatformBundle;
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
+use Symfony\Bundle\MonologBundle\MonologBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpFoundation\Session\SessionFactory;
@@ -20,6 +21,7 @@ class AppKernel extends Kernel
         $bundles = [
             new ApiPlatformBundle(),
             new DoctrineBundle(),
+            new MonologBundle(),
             new FrameworkBundle(),
         ];